Created a class to reuse the Project record
